+++ bindgen-cli/options.rs
@@ -1,6 +1,6 @@
 use bindgen::callbacks::TypeKind;
 use bindgen::{
-    builder, AliasVariation, Builder, CodegenConfig, EnumVariation,
+    builder, AliasVariation, Builder, CodegenConfig, EnumVariation, Formatter,
     FieldVisibilityKind, MacroTypeVariation, NonCopyUnionStyle, RegexSet,
     RustTarget, DEFAULT_ANON_FIELDS_PREFIX, RUST_TARGET_STRINGS,
 };
@@ -266,11 +266,11 @@ struct BindgenCommand {
     /// Do not bind size_t as usize (useful on platforms where those types are incompatible).
     #[arg(long = "no-size_t-is-usize")]
     no_size_t_is_usize: bool,
-    /// Do not format the generated bindings with rustfmt.
-    #[arg(long)]
-    no_rustfmt_bindings: bool,
-    /// The absolute path to the rustfmt configuration file. The configuration file will be used for formatting the bindings. This parameter is incompatible with --no-rustfmt-bindings.
-    #[arg(long, value_name = "PATH", conflicts_with("no_rustfmt_bindings"))]
+    /// Which tool should be used to format the bindings
+    #[arg(long, value_name = "FORMATTER", default_value="rustfmt")]
+    formatter: Formatter,
+    /// The absolute path to the rustfmt configuration file. The configuration file will be used for formatting the bindings. This parameter sets `formatter` to `rustfmt`.
+    #[arg(long, value_name = "PATH")]
     rustfmt_configuration_file: Option<String>,
     /// Avoid deriving PartialEq for types matching <REGEX>.
     #[arg(long, value_name = "REGEX")]
@@ -456,7 +456,7 @@ where
         dump_preprocessed_input,
         no_record_matches,
         no_size_t_is_usize,
-        no_rustfmt_bindings,
+        formatter,
         rustfmt_configuration_file,
         no_partialeq,
         no_copy,
@@ -821,9 +821,7 @@ where
         builder = builder.size_t_is_usize(false);
     }
 
-    if no_rustfmt_bindings {
-        builder = builder.rustfmt_bindings(false);
-    }
+    builder = builder.formatter(formatter);
 
     if let Some(path_str) = rustfmt_configuration_file {
         let path = PathBuf::from(path_str);